+++ packages/sdk-react-hooks/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @ethereum-tag-service/sdk-react-hooks
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethereum-tag-service/sdk-core@0.0.10
+
+## 0.0.9
+
+### Patch Changes
+
+- [#415](https://github.com/ethereum-tag-service/ets/pull/415) [`7299be9e4203ac122cb69d942639c4d364810305`](https://github.com/ethereum-tag-service/ets/commit/7299be9e4203ac122cb69d942639c4d364810305) Thanks [@mhpaler](https://github.com/mhpaler)! - Add README to sdk-core
+
+- Updated dependencies [[`7299be9e4203ac122cb69d942639c4d364810305`](https://github.com/ethereum-tag-service/ets/commit/7299be9e4203ac122cb69d942639c4d364810305)]:
+  - @ethereum-tag-service/sdk-core@0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`a68f6a7`](https://github.com/ethereum-tag-service/ets/commit/a68f6a70a1616c1e75e5df6e2ee4e3c546620d82)]:
+  - @ethereum-tag-service/sdk-core@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`32cd824`](https://github.com/ethereum-tag-service/ets/commit/32cd82415e907e6d143281e2d9b5436d5a44fb1c)]:
+  - @ethereum-tag-service/sdk-core@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes